refactor: migrado api/controllers/ para api/routers
