+++ src/adapter/driver/api/routers/usuarioRouter.ts
@@ -1,17 +1,16 @@
 import express from "express";
+import { Request, Response } from "express";
+import { UsuarioController } from "interfaces/controllers/usuarioController";
 
 import DBUsuariosRepository from "~adapter/driven/database/repository/usuarioDatabaseRepository";
 
-import UsuarioAPIController from "../controllers/usuarioController";
-
 import { ListaPagamentosSchema } from "./schemas/pagamentoRouter.schema";
-import { CriaUsuarioSchema, RetornaUsuarioSchema } from "./schemas/usuarioRouter.schema";
+import { CriaUsuarioBody, CriaUsuarioSchema, RetornaUsuarioBody, RetornaUsuarioSchema } from "./schemas/usuarioRouter.schema";
 import { validaRequisicao } from "./utils";
 
 const usuarioRouter = express.Router();
 
-const dbUsuariosRepository = new DBUsuariosRepository();
-const usuarioAPIController = new UsuarioAPIController(dbUsuariosRepository);
+const dbUsuarioRepository = new DBUsuariosRepository();
 
 /** 
  * @openapi
@@ -98,7 +97,32 @@ const usuarioAPIController = new UsuarioAPIController(dbUsuariosRepository);
  */
 usuarioRouter.post("/",
   validaRequisicao(CriaUsuarioSchema),
-  usuarioAPIController.criaUsuario.bind(usuarioAPIController)
+  async (
+    req: Request<unknown, CriaUsuarioBody>,
+    res: Response
+  ) => {
+    try {
+      const usuario = req.body;
+
+      const usuarioCriado = await UsuarioController.criaUsuario(dbUsuarioRepository, usuario);
+      return res.status(201).json({
+        status: "success",
+        message: usuarioCriado,
+      });
+    } catch (err: any) {
+      if (err.message === "usuario_duplicado") {
+        return res.status(400).json({
+          status: "error",
+          message: "Email ou CPF já em uso!"
+        })
+      }
+      return res.status(500).json({
+        status: "error",
+        message: err.message,
+      });
+    }
+  }
+
 );
 
 /**
@@ -128,7 +152,24 @@ usuarioRouter.post("/",
  */
 usuarioRouter.get("/",
   validaRequisicao(ListaPagamentosSchema),
-  usuarioAPIController.listaUsuarios.bind(usuarioAPIController)
+  async (
+    req: Request,
+    res: Response
+  ) => {
+    try {
+      const usuarios = await UsuarioController.listaUsuarios(dbUsuarioRepository);
+
+      return res.status(200).json({
+        status: "success",
+        usuarios,
+      });
+    } catch (err: any) {
+      return res.status(500).json({
+        status: "error",
+        message: err.message,
+      });
+    }
+  }
 );
 
 /**
@@ -180,7 +221,33 @@ usuarioRouter.get("/",
  */
 usuarioRouter.post("/query",
   validaRequisicao(RetornaUsuarioSchema),
-  usuarioAPIController.retornaUsuario.bind(usuarioAPIController)
+  async (
+    req: Request<unknown, RetornaUsuarioBody>,
+    res: Response
+  ) => {
+    try {
+      const { body } = req;
+
+      const usuario = await UsuarioController.retornaUsuario(dbUsuarioRepository, body.cpf);
+
+      if (usuario) {
+        return res.status(200).json({
+          status: "success",
+          usuario,
+        });
+      }
+      return res.status(404).json({
+        status: "error",
+        message: "Usuário não encontrado!",
+      });
+
+    } catch (err: any) {
+      return res.status(500).json({
+        status: "error",
+        message: err.message,
+      });
+    }
+  }
 );
 
 export default usuarioRouter;
